添加渠道子类接口

+++ common/src/main/java/com/godfunc/entity/MerchantChannelRate.java
@@ -37,7 +37,7 @@ public class MerchantChannelRate implements Serializable {
     /**
      * 费率
      */
-    private Double rate;
+    private Float rate;
 
     /**
      * 创建人id
